refactor: change ProductService export
