+++ src/adapters/inbound/factories/create-product-controller-factory.ts
@@ -1,4 +1,4 @@
-import ProductService from '@application/services/product-service';
+import { ProductService } from '@application/services';
 import ProductRepository from '@adapters/outbound/persistence/typeorm/product-repository';
 import { CreateProductController } from '@adapters/inbound/controllers';
 import { Factory } from '@adapters/inbound/interfaces';
