+++ src/adapters/inbound/tests/controllers/get-products-controller.spec.ts
@@ -6,7 +6,7 @@ import ProductRepository from '@adapters/outbound/persistence/typeorm/product-re
 import Product from '@application/domain/product';
 import ProductRepositoryPort from '@application/ports/product-repository-port';
 import ProductServicePort from '@application/ports/product-service-port';
-import ProductService from '@application/services/product-service';
+import { ProductService } from '@application/services';
 
 describe('Get Product Controller', () => {
   let sysUnderTest: ControllerInterface;
